Update for default mode
- Extend wait time for color wipe from 8 bit to 16 bit
- Add default boring normal mode, delay may be having some problems

+++ helpers.ino
@@ -1,6 +1,6 @@
 
 // Fill the dots one after the other with a color
-void colorWipe(uint32_t c, uint8_t wait) {
+void colorWipe(uint32_t c, uint16_t wait) {
   for(uint16_t i=0; i<strip.numPixels(); i++) {
       strip.setPixelColor(i, c);
       strip.show();
